Add fixes

+++ app/src/main/java/com/example/homeworks/responce/Wind.kt
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName
 
 data class Wind(
     @SerializedName("deg")
-    var deg: Int,
+    var deg: Double,
     @SerializedName("gust")
-    var gust: Int,
+    var gust: Double,
     @SerializedName("speed")
-    var speed: Int
+    var speed: Double
 )
